+++ ttnhabbridge/src/main/java/nl/sikken/bertrik/hab/habitat/docs/ListenerTelemetryDoc.java
@@ -5,7 +5,7 @@ import java.time.Instant;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import nl.sikken.bertrik.hab.habitat.HabReceiver;
+import nl.sikken.bertrik.hab.HabReceiver;
 
 /**
  * Listener telemetry doc.
